net: drop_dump: Fix pointer-to-int-cast errors
../net/core/drop_dump.c:168:16: error: cast to smaller integer type 'unsigned int' from 'unsigned char *' [-Werror,-Wpointer-to-int-cast]
                                push_len = (unsigned int)skb2->data
                                           ^~~~~~~~~~~~~~~~~~~~~~~~
../net/core/drop_dump.c:169:11: error: cast to smaller integer type 'unsigned int' from 'unsigned char *' [-Werror,-Wpointer-to-int-cast]
                                           - (unsigned int)skb_network_header(skb2);
                                             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
2 errors generated.

+++ net/core/drop_dump.c
@@ -162,11 +162,11 @@ queueing:
 				push_len = (unsigned int)skb_network_header_len(skb2);
 			} else if (skb_headroom(skb2) > skb2->transport_header) {
 				struct tcphdr *tcph = tcp_hdr(skb2);
-				push_len = (unsigned int)skb_network_header_len(skb2) 
+				push_len = (unsigned int)skb_network_header_len(skb2)
 					   + (unsigned int)(tcph->doff * 4);
 			} else {
-				push_len = (unsigned int)skb2->data 
-					   - (unsigned int)skb_network_header(skb2);
+				push_len = (uintptr_t)skb2->data
+					   - (uintptr_t)skb_network_header(skb2);
 			}
 
 			if (unlikely(skb_headroom(skb2) < push_len)) {
